Introduce scriptlet
Shell module re design, global shell utilization
