+++ cbok/__init__.py
@@ -1,6 +1,10 @@
 __version__ = "0.3"
 
-import pymysql
+try:
+    import pymysql
+except ModuleNotFoundError:
+    pymysql = None
 
-pymysql.version_info = (2, 2, 1, "final", 0)
-pymysql.install_as_MySQLdb()
+if pymysql is not None:
+    pymysql.version_info = (2, 2, 1, "final", 0)
+    pymysql.install_as_MySQLdb()
