+++ cbok/bbx/put_patch/worker/diff.sh
@@ -4,8 +4,7 @@ set -ex
 
 
 base_path=$(python -c "from cbok import settings; print(settings.BASE_DIR)")
-source "$base_path/utils.sh"
-source "$base_path/cbok/bbx/utils.sh"
+source "$base_path/scriptlet/bootstrap.sh"
 
 workspace=$(python -c "from cbok import settings; print(settings.Workspace)")
