+++ cbok/bbx/put_patch/worker/pod_daemonset.sh
@@ -2,7 +2,7 @@
 set -ex
 
 base_path=$(python -c "from cbok import settings; print(settings.BASE_DIR)")
-source "$base_path/utils.sh"
+source "$base_path/scriptlet/bootstrap.sh"
 
 
 function init_pod(){
@@ -36,18 +36,18 @@ function init_pod(){
         kubectl apply -f /tmp/$replica
     "
 
-    num=$(gtimeout -s KILL 10 ssh root@$address "
+    num=$(cbok_timeout 10 ssh root@$address "
         kubectl get pod -n openstack -o wide -l component=compute,application=nova --no-headers | grep $node | wc -l
     ")
     if [[ ! "$num" -eq 1 ]]; then
         die "too many replica left: $replica"
     fi
 
-    pod_name=$(gtimeout -s KILL 10 ssh root@$address "
+    pod_name=$(cbok_timeout 10 ssh root@$address "
         kubectl get pod -n openstack -o wide -l component=compute,application=nova --no-headers | grep $node | awk '{print \$1}'
     ")
 
-    gtimeout -s KILL 10 ssh root@$address "
+    cbok_timeout 10 ssh root@$address "
         kubectl delete pod -n openstack "$pod_name" --force
     "
 
@@ -73,14 +73,6 @@ function init_pod(){
         fi
         sleep 1
     done
-}
-
-
-function remote_exec_via_jump() {
-    local jump=$1
-    shift
-    ssh_key="ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3"
-    sshpass -p "easystack" ssh root@"$jump" $ssh_key $@
 }
 
 
@@ -113,23 +105,23 @@ sed -i "/^  $(basename $start_script): |\$/{
 kubectl apply -f /tmp/$replica
 EOF
 
-    num=$(gtimeout -s KILL 10 sshpass -p "easystack" ssh root@$address "ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3 \"
+    num=$(cbok_timeout 10 sshpass -p "easystack" ssh root@$address "ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3 \"
     kubectl get pod -n openstack -o wide -l component=compute,application=nova --no-headers | grep $node | wc -l
     \"")
     if [[ ! "$num" -eq 1 ]]; then
         die "too many replica left: $replica"
     fi
 
-    pod_entry=$(gtimeout -s KILL 10 sshpass -p "easystack" ssh root@$address "ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3 \"
+    pod_entry=$(cbok_timeout 10 sshpass -p "easystack" ssh root@$address "ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3 \"
         kubectl get pod -n openstack -o wide -l component=compute,application=nova --no-headers | grep $node
     \"")
     pod_name=$(echo "$pod_entry" | awk '{print $1}')
-    gtimeout -s KILL 10 sshpass -p "easystack" ssh root@$address "ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3 \"
+    cbok_timeout 10 sshpass -p "easystack" ssh root@$address "ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3 \"
         kubectl delete pod -n openstack "$pod_name" --force
     \""
 
     for i in {1..30}; do
-        restarted=$(gtimeout -s KILL 10 sshpass -p "easystack" ssh root@$address "ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3 \"
+        restarted=$(cbok_timeout 10 sshpass -p "easystack" ssh root@$address "ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3 \"
             kubectl get pod -o wide -n openstack | grep $replica | grep $node || true
         \"")
         running=$(echo "$restarted" | grep PodInitializing || true)
