+++ cbok/bbx/put_patch/worker/pod_deployment.sh
@@ -2,7 +2,7 @@
 set -ex
 
 base_path=$(python -c "from cbok import settings; print(settings.BASE_DIR)")
-source "$base_path/utils.sh"
+source "$base_path/scriptlet/bootstrap.sh"
 
 
 function init_pod(){
@@ -11,7 +11,7 @@ function init_pod(){
     replica=$3
     configmap=$4
     start_script=$5
-    gtimeout -s KILL 10 ssh root@$address "
+    cbok_timeout 10 ssh root@$address "
         kubectl scale $controller/$replica --replicas=1 -n openstack
     "
     ssh root@$address "
@@ -45,7 +45,7 @@ function init_pod(){
         grep_flag="$replica"
     fi
     for i in {1..30}; do
-        num=$(gtimeout -s KILL 10 ssh root@$address "
+        num=$(cbok_timeout 10 ssh root@$address "
             kubectl get pod -n openstack -l component=$grep_flag --field-selector=status.phase=Running --no-headers | wc -l
         ")
         if [[ "$num" -eq 1 ]]; then
@@ -58,15 +58,15 @@ function init_pod(){
         sleep 1
     done
 
-    pod_name=$(gtimeout -s KILL 10 ssh root@$address "
+    pod_name=$(cbok_timeout 10 ssh root@$address "
         kubectl get pod -n openstack | grep $replica | awk '{print \$1}'
     ")
-    gtimeout -s KILL 10 ssh root@$address "
+    cbok_timeout 10 ssh root@$address "
         kubectl delete pod -n openstack "$pod_name"
     "
 
     for i in {1..30}; do
-        restarted=$(gtimeout -s KILL 10 ssh root@$address "
+        restarted=$(cbok_timeout 10 ssh root@$address "
             kubectl get pod -n openstack | grep $replica || true
         ")
         running=$(echo "$restarted" | grep Running || true)
@@ -90,21 +90,13 @@ function init_pod(){
 }
 
 
-function remote_exec_via_jump() {
-    local jump=$1
-    shift
-    ssh_key="ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3"
-    sshpass -p "easystack" ssh root@"$jump" $ssh_key $@
-}
-
-
 function init_pod_os_in_os(){
     address=$1
     controller=$2
     replica=$3
     configmap=$4
     start_script=$5
-    gtimeout -s KILL 10 sshpass -p "easystack" ssh root@$address "ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3 \"
+    cbok_timeout 10 sshpass -p "easystack" ssh root@$address "ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3 \"
         kubectl scale $controller/$replica --replicas=1 -n openstack
     \""
     remote_exec_via_jump "$address" bash -s <<EOF
@@ -134,7 +126,7 @@ EOF
         grep_flag="$replica"
     fi
     for i in {1..30}; do
-        num=$(gtimeout -s KILL 10 sshpass -p "easystack" ssh root@$address "ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3 \"
+        num=$(cbok_timeout 10 sshpass -p "easystack" ssh root@$address "ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3 \"
         kubectl get pod -n openstack -l component=$grep_flag --field-selector=status.phase=Running --no-headers | wc -l
         \"")
         if [[ "$num" -eq 1 ]]; then
@@ -147,16 +139,16 @@ EOF
         sleep 1
     done
 
-    pod_entry=$(gtimeout -s KILL 10 sshpass -p "easystack" ssh root@$address "ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3 \"
+    pod_entry=$(cbok_timeout 10 sshpass -p "easystack" ssh root@$address "ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3 \"
         kubectl get pod -n openstack | grep $replica
     \"")
     pod_name=$(echo "$pod_entry" | awk '{print $1}')
-    gtimeout -s KILL 10 sshpass -p "easystack" ssh root@$address "ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3 \"
+    cbok_timeout 10 sshpass -p "easystack" ssh root@$address "ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3 \"
         kubectl delete pod -n openstack "$pod_name"
     \""
 
     for i in {1..30}; do
-        restarted=$(gtimeout -s KILL 10 sshpass -p "easystack" ssh root@$address "ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3 \"
+        restarted=$(cbok_timeout 10 sshpass -p "easystack" ssh root@$address "ssh -i ~/.ssh/id_rsa.roller root@10.20.0.3 \"
             kubectl get pod -n openstack | grep $replica || true
         \"")
         running=$(echo "$restarted" | grep Running || true)
