+++ cbok/bbx/virt_devel/build-virt.sh
@@ -5,7 +5,7 @@
 set -ex
 
 base_path=$(python -c "from cbok import settings; print(settings.BASE_DIR)")
-source "$base_path/utils.sh"
+source "$base_path/scriptlet/bootstrap.sh"
 
 distro "CentOS 7.9" "3.6"
