+++ cbok/cmd/__main__.py
@@ -5,5 +5,5 @@ if __name__ == "__main__":
         sys.argv = ["cbok"] + sys.argv[3:]
     else:
         sys.argv = ["cbok"] + sys.argv[1:]
-    from cbok.cmd.base import main
+    from cbok.cmd.main import main
     sys.exit(main() or 0)
